feat: simplify impl
