+++ llm-chain-llama/src/lib.rs
@@ -23,7 +23,6 @@
 
 mod context;
 mod executor;
-mod stream_executor;
 mod instruct;
 mod output;
 mod step;
@@ -31,7 +30,6 @@ mod tokenizer;
 
 pub use context::LlamaContextParams;
 pub use executor::Executor;
-pub use stream_executor::StreamExecutor;
 pub use instruct::new_instruct_template;
 pub use output::Output;
 pub use step::{LlamaConfig, Step};
